+++ backend/controllers/product.controller.js
@@ -5,10 +5,10 @@ import Product from "../models/product.model.js"
 export const getProducts = async (req, res) => {
   try {
     const products = await Product.find({})
-    res.status(200).json({ success: true, data: products })
+    res.status(200).json(products)
   } catch (error) {
-    console.log("error in fetching products:", error.message)
-    res.status(500).json({ success: false, message: "Server Error" })
+    console.log("error in fetching products:", error.msg)
+    res.status(500).json({ msg: "Server Error" })
   }
 }
 
@@ -17,9 +17,7 @@ export const createProduct = async (req, res) => {
   const product = req.body
 
   if (!product.name || !product.price || !product.imageUrl) {
-    return res
-      .status(400)
-      .json({ success: false, message: "Please provide all fields" })
+    return res.status(400).json({ msg: "Please provide all fields" })
   }
 
   const newProduct = new Product(product)
@@ -28,8 +26,8 @@ export const createProduct = async (req, res) => {
     await newProduct.save()
     res.status(201).json({ success: true, data: newProduct })
   } catch (error) {
-    console.error("Error in Create product:", error.message)
-    res.status(500).json({ success: false, message: "Server Error" })
+    console.error("Error in Create product:", error.msg)
+    res.status(500).json({ msg: "Server Error" })
   }
 }
 
@@ -39,18 +37,16 @@ export const updateProduct = async (req, res) => {
   const product = req.body
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
-    return res
-      .status(404)
-      .json({ success: false, message: "Invalid Product Id" })
+    return res.status(404).json({ msg: "Invalid Product Id" })
   }
 
   try {
     const updatedProduct = await Product.findByIdAndUpdate(id, product, {
       new: true,
     })
-    res.status(200).json({ success: true, data: updatedProduct })
+    res.status(200).json({ data: updatedProduct })
   } catch (error) {
-    res.status(500).json({ success: false, message: "Server Error" })
+    res.status(500).json({ msg: "Server Error" })
   }
 }
 
@@ -59,17 +55,15 @@ export const deleteProduct = async (req, res) => {
   const { id } = req.params
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
-    return res
-      .status(404)
-      .json({ success: false, message: "Invalid Product Id" })
+    return res.status(404).json({ msg: "Invalid Product Id" })
   }
 
   try {
-    await Product.findByIdAndDelete(id)
-    res.status(200).json({ success: true, message: "Product deleted" })
+    await Product.findByIdAndDelefte(id)
+    res.status(200).json({ msg: "Product deleted" })
   } catch (error) {
-    console.log("error in deleting product:", error.message)
-    res.status(500).json({ success: false, message: "Server Error" })
+    console.log("error in deleting product:", error.msg)
+    res.status(500).json({ msg: "Server Error" })
   }
 }
 
@@ -77,9 +71,9 @@ export const deleteProduct = async (req, res) => {
 export const deleteAllProducts = async (req, res) => {
   try {
     await Product.deleteMany({})
-    res.status(200).json({ success: true, message: "All products deleted" })
+    res.status(200).json({ msg: "All products deleted" })
   } catch (error) {
-    console.log("error in deleting all products:", error.message)
-    res.status(500).json({ success: false, message: "Server Error" })
+    console.log("error in deleting all products:", error.msg)
+    res.status(500).json({ msg: "Server Error" })
   }
 }