+++ backend/models/user.model.js
@@ -7,7 +7,7 @@ const userSchema = new mongoose.Schema({
   password: { type: String, required: true },
   address: { type: String, required: true },
   phone: { type: String, required: true },
-  isAdmin: { type: Boolean, required: true, default: false },
+  isAdmin: { type: Boolean, default: false },
 })
 
 export default mongoose.model("User", userSchema)